Corrected a mistake i discovered in the code, AnswerQuestionFrame default close operation should be DISPOSE_ON_CLOSE so doesn't close whole program when closed

+++ Quiz/src/server/QuizServer.java
@@ -143,7 +143,7 @@ public class QuizServer extends UnicastRemoteObject implements QuizRemoteInterfa
 	 * Adds the id of a quiz to list of quizzes currently being played
 	 */
 	public void addCurrentlyBeingPlayedQuiz(Integer id){
-		this.quizzesCurrentlyBeingPlayed.add(id);
+		this.quizzesCurrentlyBeingPlayed.add((Integer) id);
 		System.out.println("Added quiz + " + id + " to currently being played quizlist");
 		for (Integer ids: this.quizzesCurrentlyBeingPlayed){
 			System.out.println("ids " + ids);
@@ -153,7 +153,7 @@ public class QuizServer extends UnicastRemoteObject implements QuizRemoteInterfa
 	 * removes an id for a quiz from the list of quizzes currently being played once a user has finished playing
 	 */
 	public void removeCurrentlyBeingPlayedQuiz(Integer id){
-		this.quizzesCurrentlyBeingPlayed.remove(id);
+		this.quizzesCurrentlyBeingPlayed.remove((Integer)id);
 		System.out.println("Removed quiz + " + id + " to currently being played quizlist");
 		for (Integer ids: this.quizzesCurrentlyBeingPlayed){
 			System.out.println("ids " + ids);
